+++ src/notionDatabase.ts
@@ -1,31 +1,31 @@
 import fetch from "node-fetch";
 import NotionCreateField, { IData } from "./notionCreateField";
-import NotionField from "./notionField";
-interface INotionDatabase {
+import NotionField, { IResultProps } from "./notionField";
+interface INotionDatabaseProps {
   secretKey: string;
   databaseId: string;
   notionVersion?: string;
 }
-interface IQuery {
+interface IQueryProps {
   filter?: { [key: string]: any };
   sorts?: { [key: string]: any }[];
 }
-interface IUpdate {
+interface IUpdateProps<T> {
   pageId: string;
-  updateDatas: IData[];
+  updateDatas: IData<T>[];
 }
-export default class NotionDatabase {
-  private BASE_URL = "https://api.notion.com/v1";
-  private secretKey: string;
-  private databaseId: string;
-  private notionVersion: string;
-  private notionCreateField: NotionCreateField;
-  private notionField: NotionField;
+export default class NotionDatabase<T extends {}> {
+  private readonly BASE_URL = "https://api.notion.com/v1";
+  private readonly secretKey: string;
+  private readonly databaseId: string;
+  private readonly notionVersion: string;
+  private readonly notionCreateField: NotionCreateField<T>;
+  private readonly notionField: NotionField;
   constructor({
     secretKey,
     databaseId,
     notionVersion = "2021-05-13",
-  }: INotionDatabase) {
+  }: INotionDatabaseProps) {
     this.secretKey = secretKey;
     this.databaseId = databaseId;
     this.notionVersion = notionVersion;
@@ -41,22 +41,22 @@ export default class NotionDatabase {
     };
   }
 
-  async query({ filter, sorts }: IQuery) {
+  async query(props?: IQueryProps): Promise<IResultProps<T>[]> {
     return fetch(`${this.BASE_URL}/databases/${this.databaseId}/query`, {
       method: "POST",
       headers: this.createHeaders(),
       body: JSON.stringify({
-        filter: filter && {
-          ...filter,
+        filter: props?.filter && {
+          ...props.filter,
         },
-        sorts: sorts && [...sorts],
+        sorts: props?.sorts && [...props.sorts],
       }),
     })
       .then((res) => res.json())
-      .then((json) => this.notionField.getRows(json.results));
+      .then((json) => this.notionField.getRows<T>(json.results));
   }
 
-  async create(newDatas: IData[]) {
+  async create(newDatas: IData<T>[]): Promise<IResultProps<T>> {
     return fetch(`${this.BASE_URL}/pages/`, {
       method: "POST",
       headers: this.createHeaders(),
@@ -68,10 +68,13 @@ export default class NotionDatabase {
       }),
     })
       .then((res) => res.json())
-      .then((json) => this.notionField.getRow(json));
+      .then((json) => this.notionField.getRow<T>(json));
   }
 
-  async update({ pageId, updateDatas }: IUpdate) {
+  async update({
+    pageId,
+    updateDatas,
+  }: IUpdateProps<T>): Promise<IResultProps<T>> {
     return fetch(`${this.BASE_URL}/pages/${pageId}`, {
       method: "PATCH",
       headers: this.createHeaders(),
@@ -80,6 +83,6 @@ export default class NotionDatabase {
       }),
     })
       .then((res) => res.json())
-      .then((json) => this.notionField.getRow(json));
+      .then((json) => this.notionField.getRow<T>(json));
   }
 }