+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## node-notion-database
+
+### 0.0.4 (2023-06-28)
+
+수정 및 기능추가:
+
+- 수정: README.md 수정, notionDatabase properties 타입수정
+- 기능추가:
+
 ### 0.0.3 (2022-08-09)
 
 수정 및 기능추가: